05_redis/users_service/api/token.py
from datetime import datetime, timedelta

from jose import jwt
from fastapi import APIRouter, Depends, HTTPException, status
from fastapi.security import OAuth2PasswordRequestForm
from sqlalchemy.orm import Session

import constants.auth as auth_constants
from models.token import Token
from db.pg import get_db
from db import crud, models

router = APIRouter(tags=["auth"])

@router.post("/token", response_model=Token)
async def login_for_access_token(
    form_data: OAuth2PasswordRequestForm = Depends(),
    db: Session = Depends(get_db)
):
    user = crud.get_user_by_username(db, username=form_data.username)
    if not user or not models.User.verify_password(form_data.password, user.hashed_password):
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Incorrect username or password",
            headers={"WWW-Authenticate": "Bearer"},
        )

    access_token_expires = timedelta(minutes=auth_constants.ACCESS_TOKEN_EXPIRE_MINUTES)
    
    access_token = jwt.encode(
        {"sub": user.username, "exp": datetime.utcnow() + access_token_expires},
        auth_constants.SECRET_KEY,
        algorithm=auth_constants.ALGORITHM,
    )
    
    return {"access_token": access_token, "token_type": "bearer"}
